added a step of adding url for recognition provider
