+++ backend_jobs/recognition_pipeline/pipeline_lib/image_recognition_provider.py
@@ -16,6 +16,11 @@
 from abc import ABC, abstractmethod
 import apache_beam as beam
 from backend_jobs.recognition_pipeline.filters import filter_by_format, filter_by_resolution
+from backend_jobs.ingestion_pipeline.providers.providers import\
+  get_provider, IMAGE_PROVIDERS
+from backend_jobs.pipeline_utils import database_schema
+from backend_jobs.recognition_pipeline.pipeline_lib import constants
+
 
 class ImageRecognitionProvider(ABC, beam.DoFn):
     """ Each recognition provider used in our project
@@ -60,8 +65,50 @@ class ImageRecognitionProvider(ABC, beam.DoFn):
         """
         resolution_filter = filter_by_resolution.FilterByResolution(self._resolution_prerequisites)
         format_filter = filter_by_format.FilterByFormat(self._format_prerequisites)
-        resolution_filter.add_url_for_recognition_api(image)
         return resolution_filter.is_supported(image) and format_filter.is_supported(image)
+
+    def _change_url_by_resolution(self, image):
+        """ Adds an image's url field to match the supported resolution.
+
+            Iterates through all image providers and checks if
+            the image can be resized to a supported resolution.
+            If it can, adds a field of an image url which contains
+            an image of the supported resolution.
+
+        """
+        for provider_name in image[database_schema.COLLECTION_IMAGES_FIELD_INGESTED_PROVIDERS]:
+            provider = get_provider(IMAGE_PROVIDERS, provider_name)
+            resize_url = provider.get_url_for_min_resolution(\
+                self._resolution_prerequisites['height'],\
+                    self._resolution_prerequisites['width'], image)
+            if resize_url:
+                image[constants.FIELD_URL_FOR_RECOGNITION_API] =\
+                    resize_url
+                return image
+            return image
+
+    def add_url_for_recognition_api(self, image):
+        """ Adds a URL_FOR_RECOGNITION_API field with the url in the supported
+            resolution if possible.
+
+            If the image is in the supported resolution, copies the url to the
+            a new URL_FOR_RECOGNITION_API field. Otherwise - checks if it can be
+            resized to the correct resolution.
+            If the image cannot be supported - the dictionary remains without
+            the new field.
+
+        """
+        image_attribute = image[database_schema.COLLECTION_IMAGES_FIELD_IMAGE_ATTRIBUTES]\
+            [database_schema.COLLECTION_IMAGES_FIELD_RESOLUTION]
+        min_height = self._resolution_prerequisites['height']
+        min_width = self._resolution_prerequisites['width']
+        if image_attribute['width'] >= min_width and \
+            image_attribute['height'] >= min_height:
+            # current url is in a supported resolution
+            image[constants.FIELD_URL_FOR_RECOGNITION_API] =\
+                image[database_schema.COLLECTION_IMAGES_FIELD_URL]
+            return [image]
+        return [self._change_url_by_resolution(image)]
 
     @property
     def _resolution_prerequisites(self):
